Update second.rs
Probably no need to return `Result` only to immediately `unwrap()` it.

Another note, when `if row[x] {` is replaced to `if row[x] == '#' {` you can save initial `map` over each like of the input file. Although it could be a matter of taste and (I just realized) you ARE saving on memory usage - `bool` is 1 byte, when `char` is 4 bytes due to Unicode. So, theoretically 1 byte the comparison is faster as well. Eh... just a thought. :) I didn't bother converting the input to bool in my solution.